backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "supabase>=2.28.0",
    "uvicorn[standard]>=0.41.0",
]
